metadata file upload size and type checks added + fixed URL in metadata
Set metadata file upload limit to 20MB and added restrictions to
accepted filetypes. A hint was added to file upload button to
inform the user about said restrictions.

Also added a small fix to protocols being used in metadata URLs.
Bug here was being caused by a workaround of the situation when
you don't specify a protocol in a link being put into href.
This would result in a relative link. Therefore a check for
protocol is added and HTTPS is prepended if none is found.